% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bootstrap_effective_volume}
\alias{bootstrap_effective_volume}
\title{Bin the data for fitting the model.  This involves
estimation the volume for each range bin as well as how many
animals are in the bin. The resulting density is used to estimate steady state local density, which is then used to compute the number
of animals that are expected in each range interval. The expected / observed is the basis for determining probability of detection.}
\usage{
bootstrap_effective_volume(
  target_ranges,
  vol_func,
  nbins = 25,
  method = "median",
  nvals = 5,
  loc_dens = NULL,
  model_method = "logistic glm",
  formula = NULL,
  boot_n = 500,
  plotting = FALSE
)
}
\arguments{
\item{vol_func}{Volume function as returned by get_vol_func}

\item{nbins}{number of bins to use for density histogram}

\item{method}{method to determine the local density based on change in density by range.  Valid choices are 'median' and 'mean'}

\item{nvals}{an integer specifying how many of the highest density values to average for local density.}

\item{loc_dens}{a value that can be entered to provide a separate externally estimated local density. Default is NULL.}

\item{model_method}{flag specifying which model to use. Valid choices include 'logistic glm', ADD}

\item{formula}{optional formula to pass to the model, otherwise defaults are used}

\item{boot_n}{number of bootstrap iterations}

\item{plotting}{boolean flag to indicate whether to make plots}

\item{targets}{vector of ranges for a given target}
}
\value{
A data frame containing range, number observed and number expected. The number observed is constrained to be <=number expected.
}
\description{
Bin the data for fitting the model.  This involves
estimation the volume for each range bin as well as how many
animals are in the bin. The resulting density is used to estimate steady state local density, which is then used to compute the number
of animals that are expected in each range interval. The expected / observed is the basis for determining probability of detection.
}
