% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{detect_single_logistic}
\alias{detect_single_logistic}
\title{logistic}
\usage{
detect_single_logistic(parm, vol, rng, fishcount)
}
\arguments{
\item{parm}{Vector of parameters for logistic}

\item{vol}{to do}

\item{rng}{to do}

\item{fishcount}{to do}
}
\description{
logistic
}
