% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{eval_double_logistic}
\alias{eval_double_logistic}
\title{double logistic evaluation}
\usage{
eval_double_logistic(parm, vol, rng, fishcount)
}
\arguments{
\item{parm}{Vector of parameters for double logistic scale,r50_asc,slope_asc,r50_desc,slope_desc)}

\item{vol}{vector of volumes of each range interval}

\item{rng}{vector of midpoints of each range interval}

\item{fishcount}{vector of numbers of fish in each range interval}
}
\description{
double logistic evaluation
}
