% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{eval_normal}
\alias{eval_normal}
\title{normal evaluation}
\usage{
eval_normal(parm, vol, rng, fishcount)
}
\arguments{
\item{parm}{vector of 3 parameters: scale, mean, and standard deviation of a normal curve}

\item{vol}{vector of volumes of each range interval}

\item{rng}{vector of midpoints of each range interval}

\item{fishcount}{vector of numbers of fish in each range interval}
}
\description{
normal evaluation
}
