% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{eval_single_logistic}
\alias{eval_single_logistic}
\title{single logistic evaluation}
\usage{
eval_single_logistic(parm, vol, rng, fishcount)
}
\arguments{
\item{parm}{Vector of parameters for single logistic (scale,r50,slope)}

\item{vol}{vector of volumes of each range interval}

\item{rng}{vector of midpoints of each range interval}

\item{fishcount}{vector of numbers of fish in each range interval}
}
\description{
single logistic evaluation
}
