% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_probability.R
\name{fit_density_function_covariate}
\alias{fit_density_function_covariate}
\title{density function fitting. Models specified with a method flag.}
\usage{
fit_density_function_covariate(
  range_data,
  method = c("logistic glm", "logistic gam"),
  formula = NULL,
  plotting = FALSE
)
}
\arguments{
\item{range_data}{a data frame containing range, number observed, number expected, and a covariate}

\item{method}{flag specifying which model to use. Valid choices include 'logistic glm', ADD}

\item{formula}{optional formula to pass to the model, otherwise defaults are used}

\item{plotting}{boolean flag to indicate whether to make plots}
}
\value{
A R model output structure
}
\description{
density function fitting. Models specified with a method flag.
}
