% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{func_double_logistic}
\alias{func_double_logistic}
\title{double logistic function}
\usage{
func_double_logistic(parm, x)
}
\arguments{
\item{parm}{Vector of parameters for single logistic (scale,r50,slope)}

\item{x}{vector of range values}
}
\description{
double logistic function
}
