% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{func_normal}
\alias{func_normal}
\title{normal function}
\usage{
func_normal(parm, x)
}
\arguments{
\item{parm}{Vector of parameters for single logistic (scale,r50,slope)}

\item{x}{vector of range values}
}
\description{
normal function
}
