% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{func_single_logistic}
\alias{func_single_logistic}
\title{single logistic function}
\usage{
func_single_logistic(parm, x)
}
\arguments{
\item{parm}{Vector of parameters for single logistic (scale,r50,slope)}

\item{x}{vector of range values}
}
\description{
single logistic function
}
