% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{get_dens}
\alias{get_dens}
\title{Bin the data for fitting the model.  This involves
estimation the volume for each range bin as well as how many
animals are in the bin.}
\usage{
get_dens(targets, vol_func, nbins = 25)
}
\arguments{
\item{targets}{testing}

\item{vol_func}{Volume function as returned by}

\item{nbins}{to do}
}
\value{
A vector of densities for each bin
}
\description{
Bin the data for fitting the model.  This involves
estimation the volume for each range bin as well as how many
animals are in the bin.
}
