% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pollock}
\alias{pollock}
\title{Walleye pollock data}
\format{
\subsection{\code{pollock}}{

A list with two elements: 'cal' is a list from a yaml calibration file and 'targets' data frame with 1,511 rows and 7 columns:
\describe{
\item{FRAME_NUMBER}{Frame number}
\item{HX, HY, HZ}{Positions in 3d space?}
...
}
}
}
\usage{
pollock
}
\description{
A subset of data pollock in the Eastern Bering Sea
}
\keyword{datasets}
