% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_probability.R
\name{prep_detection_data}
\alias{prep_detection_data}
\title{Bin the data for fitting the model.  This involves
estimation the volume for each range bin as well as how many
animals are in the bin. The resulting density is used to estimate steady state local density, which is then used to compute the number
of animals that are expected in each range interval. The expected / observed is the basis for determining probability of detection.}
\usage{
prep_detection_data(
  target_ranges,
  vol_func,
  nbins = 25,
  method = "median",
  nvals = 5,
  loc_dens = NULL,
  target_extent = NULL,
  plotting = FALSE
)
}
\arguments{
\item{vol_func}{Volume function as returned by get_vol_func}

\item{nbins}{number of bins to use for density histogram}

\item{method}{method to determine the local density based on change in density by range.  Valid choices are 'median' and 'mean'}

\item{nvals}{an integer specifying how many of the highest density values to average for local density.}

\item{loc_dens}{a value that can be entered to provide a separate externally estimated local density. Default is NULL.}

\item{target_extent}{range of acceptable target ranges. If NULL the max target distance is used.}

\item{plotting}{boolean flag to indicate whether to make plots}

\item{targets}{vector of ranges for a given target}
}
\value{
A list containing a data frame containing range, number observed and number expected, and the estimate of max local density. The number observed is constrained to be <=number expected.
}
\description{
Bin the data for fitting the model.  This involves
estimation the volume for each range bin as well as how many
animals are in the bin. The resulting density is used to estimate steady state local density, which is then used to compute the number
of animals that are expected in each range interval. The expected / observed is the basis for determining probability of detection.
}
