% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume.R
\name{get_vol_func}
\alias{get_vol_func}
\title{Get volume from a Matlab input list. Everything hard-coded still FIX ME}
\usage{
get_vol_func(matcal)
}
\arguments{
\item{matcal}{A list of calibration values from the \code{readMat} function.}
}
\value{
A list containing the function vol_func and a vector of coefficients p_vol
}
\description{
Get volume from a Matlab input list. Everything hard-coded still FIX ME
}
\details{
to do
}
\references{
Bouget (2008) Camera calibration toolbox for Matlab. Available from http://vision.caltech.edu/bouguetj/calib_doc/index.html (accessed September 2008)].
}
