% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume.R
\name{get_vol_func}
\alias{get_vol_func}
\title{Get volume function (change in joint imaging volume with range) from a calibration.}
\usage{
get_vol_func(Cal, max_extent = 8, grid_size = 0.1, plotting = FALSE)
}
\arguments{
\item{max_extent}{distance in m determining range extent of volume field.}

\item{grid_size}{grid spacing in m determining the density of the point cloud.}

\item{plotting}{boolean flag to indicate whether to make plots}

\item{cal}{A yaml file (full path) from which to read the calibration values.}
}
\value{
A list containing the function vol_func and a vector of coefficients p_vol
}
\description{
Get volume function (change in joint imaging volume with range) from a calibration.
}
\details{
to do
}
\references{
Bouget (2008) Camera calibration toolbox for Matlab. Available from http://vision.caltech.edu/bouguetj/calib_doc/index.html (accessed September 2008)].
}
